move gpt parts to modeling_vqbet.py and change filename of vqbet_utils.py to vqvae_utils.py
