+++ lerobot/common/policies/vqbet/vqvae_utils.py
@@ -14,24 +14,11 @@ from torch.cuda.amp import autocast
 from lerobot.common.policies.vqbet.configuration_vqbet import VQBeTConfig
 
 """
-This file is part of a VQ-BeT that utilizes code from the following repositories:
+This file is a part for Residual Vector Quantization that utilizes code from the following repository:
 
-    - The Vector Quantization part is an adaptation of Phil Wang's vector-quantize-pytorch implementation in PyTorch.
+    - Phil Wang's vector-quantize-pytorch implementation in PyTorch.
         Origianl source: https://github.com/lucidrains/vector-quantize-pytorch
     
-
-    - The nanoGPT part is an adaptation of Andrej Karpathy's nanoGPT implementation in PyTorch.
-        Original source: https://github.com/karpathy/nanoGPT
-
-"""
-
-
-"""
-Vector Quantization part
-
-From this point on, the code is an adaptation of the original code from the vector-quantize-pytorch repository.
-Origianl source: https://github.com/lucidrains/vector-quantize-pytorch
-
     - The vector-quantize-pytorch code is licensed under the MIT License:
 
         MIT License
@@ -1275,252 +1262,3 @@ class EuclideanCodebook(nn.Module):
         return quantize, embed_ind, dist
 
 
-
-"""
-nanoGPT code
-
-From this point on, the code is an adaptation of the original code from the nanoGPT repository.
-Original source: https://github.com/karpathy/nanoGPT
-
-    - The nanoGPT code is licensed under the MIT License:
-
-    MIT License
-
-    Copyright (c) 2022 Andrej Karpathy
-
-    Permission is hereby granted, free of charge, to any person obtaining a copy
-    of this software and associated documentation files (the "Software"), to deal
-    in the Software without restriction, including without limitation the rights
-    to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-    copies of the Software, and to permit persons to whom the Software is
-    furnished to do so, subject to the following conditions:
-
-    The above copyright notice and this permission notice shall be included in all
-    copies or substantial portions of the Software.
-
-    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-    SOFTWARE.
-
-    - We've made some changes to the original code to adapt it to our needs.
-
-        Changed variable names:
-            - n_head -> gpt_n_head
-            - n_embd -> gpt_hidden_dim
-            - block_size -> gpt_block_size
-            - n_layer -> gpt_n_layer
-        
-        
-        class GPT(nn.Module):
-            - removed unused functions `def generate`, `def estimate_mfu`, and `def from_pretrained`
-            - changed the `configure_optimizers` to `def configure_parameters` and made it to return only the parameters of the model: we use an external optimizer in our training loop.
-            - in the function `forward`, we removed target loss calculation parts, since it will be calculated in the training loop (after passing through bin prediction and offset prediction heads).
-        
-"""
-
-class CausalSelfAttention(nn.Module):
-    def __init__(self, config):
-        super().__init__()
-        assert config.gpt_hidden_dim % config.gpt_n_head == 0
-        # key, query, value projections for all heads, but in a batch
-        self.c_attn = nn.Linear(config.gpt_hidden_dim, 3 * config.gpt_hidden_dim)
-        # output projection
-        self.c_proj = nn.Linear(config.gpt_hidden_dim, config.gpt_hidden_dim)
-        # regularization
-        self.attn_dropout = nn.Dropout(config.dropout)
-        self.resid_dropout = nn.Dropout(config.dropout)
-        # causal mask to ensure that attention is only applied to the left in the input sequence
-        self.register_buffer(
-            "bias",
-            torch.tril(torch.ones(config.gpt_block_size, config.gpt_block_size)).view(
-                1, 1, config.gpt_block_size, config.gpt_block_size
-            ),
-        )
-        self.gpt_n_head = config.gpt_n_head
-        self.gpt_hidden_dim = config.gpt_hidden_dim
-
-    def forward(self, x):
-        (
-            B,
-            T,
-            C,
-        ) = x.size()  # batch size, sequence length, embedding dimensionality (gpt_hidden_dim)
-
-        # calculate query, key, values for all heads in batch and move head forward to be the batch dim
-        q, k, v = self.c_attn(x).split(self.gpt_hidden_dim, dim=2)
-        k = k.view(B, T, self.gpt_n_head, C // self.gpt_n_head).transpose(
-            1, 2
-        )  # (B, nh, T, hs)
-        q = q.view(B, T, self.gpt_n_head, C // self.gpt_n_head).transpose(
-            1, 2
-        )  # (B, nh, T, hs)
-        v = v.view(B, T, self.gpt_n_head, C // self.gpt_n_head).transpose(
-            1, 2
-        )  # (B, nh, T, hs)
-
-        # causal self-attention; Self-attend: (B, nh, T, hs) x (B, nh, hs, T) -> (B, nh, T, T)
-        att = (q @ k.transpose(-2, -1)) * (1.0 / math.sqrt(k.size(-1)))
-        att = att.masked_fill(self.bias[:, :, :T, :T] == 0, float("-inf"))
-        att = F.softmax(att, dim=-1)
-        att = self.attn_dropout(att)
-        y = att @ v  # (B, nh, T, T) x (B, nh, T, hs) -> (B, nh, T, hs)
-        y = (
-            y.transpose(1, 2).contiguous().view(B, T, C)
-        )  # re-assemble all head outputs side by side
-
-        # output projection
-        y = self.resid_dropout(self.c_proj(y))
-        return y
-
-
-
-class Block(nn.Module):
-    def __init__(self, config):
-        super().__init__()
-        self.ln_1 = nn.LayerNorm(config.gpt_hidden_dim)
-        self.attn = CausalSelfAttention(config)
-        self.ln_2 = nn.LayerNorm(config.gpt_hidden_dim)
-        self.mlp = nn.Sequential(
-                    nn.Linear(config.gpt_hidden_dim, 4 * config.gpt_hidden_dim),
-                    nn.GELU(),
-                    nn.Linear(4 * config.gpt_hidden_dim, config.gpt_hidden_dim),
-                    nn.Dropout(config.dropout)
-                )
-
-    def forward(self, x):
-        x = x + self.attn(self.ln_1(x))
-        x = x + self.mlp(self.ln_2(x))
-        return x
-
-
-
-
-class GPT(nn.Module):
-    """
-    Original comments:
-    Full definition of a GPT Language Model, all of it in this single file.
-    References:
-    1) the official GPT-2 TensorFlow implementation released by OpenAI:
-    https://github.com/openai/gpt-2/blob/master/src/model.py
-    2) huggingface/transformers PyTorch implementation:
-    https://github.com/huggingface/transformers/blob/main/src/transformers/models/gpt2/modeling_gpt2.py
-    """
-
-
-    def __init__(self, config: VQBeTConfig):
-        super().__init__()
-        assert config.gpt_output_dim is not None
-        assert config.gpt_block_size is not None
-        self.config = config
-
-        self.transformer = nn.ModuleDict(
-            dict(
-                wte=nn.Linear(config.gpt_input_dim, config.gpt_hidden_dim),
-                wpe=nn.Embedding(config.gpt_block_size, config.gpt_hidden_dim),
-                drop=nn.Dropout(config.dropout),
-                h=nn.ModuleList([Block(config) for _ in range(config.gpt_n_layer)]),
-                ln_f=nn.LayerNorm(config.gpt_hidden_dim),
-            )
-        )
-        self.lm_head = nn.Linear(config.gpt_hidden_dim, config.gpt_output_dim, bias=False)
-        # init all weights, and apply a special scaled init to the residual projections, per GPT-2 paper
-        self.apply(self._init_weights)
-        for pn, p in self.named_parameters():
-            if pn.endswith("c_proj.weight"):
-                torch.nn.init.normal_(
-                    p, mean=0.0, std=0.02 / math.sqrt(2 * config.gpt_n_layer)
-                )
-
-        # report number of parameters
-        n_params = sum(p.numel() for p in self.parameters())
-        print("number of parameters: %.2fM" % (n_params / 1e6,))
-
-    def forward(self, input, targets=None):
-        device = input.device
-        b, t, d = input.size()
-        assert (
-            t <= self.config.gpt_block_size
-        ), f"Cannot forward sequence of length {t}, block size is only {self.config.gpt_block_size}"
-        pos = torch.arange(0, t, dtype=torch.long, device=device).unsqueeze(
-            0
-        )  # shape (1, t)
-
-        # forward the GPT model itself
-        tok_emb = self.transformer.wte(
-            input
-        )  # token embeddings of shape (b, t, gpt_hidden_dim)
-        pos_emb = self.transformer.wpe(
-            pos
-        )  # position embeddings of shape (1, t, gpt_hidden_dim)
-        x = self.transformer.drop(tok_emb + pos_emb)
-        for block in self.transformer.h:
-            x = block(x)
-        x = self.transformer.ln_f(x)
-        logits = self.lm_head(x)
-        return logits
-
-    def _init_weights(self, module):
-        if isinstance(module, nn.Linear):
-            torch.nn.init.normal_(module.weight, mean=0.0, std=0.02)
-            if module.bias is not None:
-                torch.nn.init.zeros_(module.bias)
-        elif isinstance(module, nn.Embedding):
-            torch.nn.init.normal_(module.weight, mean=0.0, std=0.02)
-        elif isinstance(module, nn.LayerNorm):
-            torch.nn.init.zeros_(module.bias)
-            torch.nn.init.ones_(module.weight)
-
-    def crop_block_size(self, gpt_block_size):
-        assert gpt_block_size <= self.config.gpt_block_size
-        self.config.gpt_block_size = gpt_block_size
-        self.transformer.wpe.weight = nn.Parameter(
-            self.transformer.wpe.weight[:gpt_block_size]
-        )
-        for block in self.transformer.h:
-            block.attn.bias = block.attn.bias[:, :, :gpt_block_size, :gpt_block_size]
-
-    def configure_parameters(self):
-        """
-        This long function is unfortunately doing something very simple and is being very defensive:
-        We are separating out all parameters of the model into two buckets: those that will experience
-        weight decay for regularization and those that won't (biases, and layernorm/embedding weights).
-        """
-
-        # separate out all parameters to those that will and won't experience regularizing weight decay
-        decay = set()
-        no_decay = set()
-        whitelist_weight_modules = (torch.nn.Linear,)
-        blacklist_weight_modules = (torch.nn.LayerNorm, torch.nn.Embedding)
-        for mn, m in self.named_modules():
-            for pn, p in m.named_parameters():
-                fpn = "%s.%s" % (mn, pn) if mn else pn  # full param name
-                if pn.endswith("bias"):
-                    # all biases will not be decayed
-                    no_decay.add(fpn)
-                elif pn.endswith("weight") and isinstance(m, whitelist_weight_modules):
-                    # weights of whitelist modules will be weight decayed
-                    decay.add(fpn)
-                elif pn.endswith("weight") and isinstance(m, blacklist_weight_modules):
-                    # weights of blacklist modules will NOT be weight decayed
-                    no_decay.add(fpn)
-
-        # validate that we considered every parameter
-        param_dict = {pn: p for pn, p in self.named_parameters()}
-        inter_params = decay & no_decay
-        union_params = decay | no_decay
-        assert (
-            len(inter_params) == 0
-        ), "parameters %s made it into both decay/no_decay sets!" % (str(inter_params),)
-        assert (
-            len(param_dict.keys() - union_params) == 0
-        ), "parameters %s were not separated into either decay/no_decay set!" % (
-            str(param_dict.keys() - union_params),
-        )
-
-        decay = [param_dict[pn] for pn in sorted(list(decay))]
-        no_decay = [param_dict[pn] for pn in sorted(list(no_decay))]
-        return decay, no_decay
